modules/project-labels-checker/function/requirements.txt
google-api-python-client==2.118.0
functions-framework==3.5.0
google-auth==2.28.0